refactoring codes for Invoices

+++ pair2/src/main/java/com/tobeto/pair2/entitites/concretes/Invoice.java
@@ -7,8 +7,6 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDate;
-
 @Table(name = "invoices")
 @Entity
 @Getter
@@ -21,7 +19,7 @@ public class Invoice extends BaseEntity {
     private String invoiceNo;
 
     @Column(name="total_price")
-    private Float totalPrice;
+    private Double totalPrice;
 
 //    @Column(name="tax_rate")
 //    private Float taxRate;
